nWelcome: make sure the user sees the reason for a failure
If the game is closed, for example, the user should be able to see why
they can't get in.  Due to the better packet error checking in
previous commits, we need to let nWelcome handle it's own packets in
these situations, so the connection isn't dropped before the error can
be displayed.