Redo the whole game "closed" thing
In the past, there was the ability to set a game as "closed" by
setting a bool in the common block via conqoper.  This would then kick
everyone out who didn't have the PLAYWHENCLOSED privilege.

Now, we turn this into a server flags/option.  Like before, when you
set the game to closed, only existing users who have the
PLAYWHENCLOSED priv set will be able to login and play.  Any current
users in the game will be left alone.

The fact that it is a server flag now also means a user can be made
aware of a server's closed status beforehand and not have to try to
log into a server just to be told it's closed :)