Rename user/sys config files from conquestrc to conquest.conf
These files are quite different from the CQI configuration files, so
to avoid confusion, rename the system and user configuration
files to conquest.conf.

These are located in <prefix>/etc/conquest/ (for the system-wide
version) and ~/.conquest/conquest.conf for the user's version.

If you have an existing conquestrc file, rename it to conquest.conf to
preserve your settings.