Always send planet location data for a planet you are orbiting
To save on bandwidth, planet location data is only sent every 5
seconds (the client fakes the motion in the intervals between
updates).

This can cause a "jitter" on the display (viewer display jumping
around between frames) when entering the game until the planet is
updated.  It can also happen at anytime your ships location diverges
much from the location of the planet you are orbiting.

Now, when your ship is orbiting a planet, it's location data is always
sent regardless of the update timer.  So far, the jitters seem to have
gone away.  Time will tell.