Do not reset transmit timer for planet location recording packets
When doing a server recording, the planet location packet timer was
being reset when such a packet was being written to the recording
file.  This caused the possibility of large delays in planet location
updates being sent to a client (outside the 5 second limit).  Once
that happened, the jitters would occur.

Now we do not reset this timer for recorded packets, only for packets
sent.  Hopefully this will be the last of the "jitter" problem.