Don't adjust ship location timer when recording
This issue didn't cause jitter, but could cause excessive delays when
updating a ships position if a server recording was in progress.  The
idea here was to ensure that a ship location packet for your ship was
always sent every 5 seconds when UDP was in use.  Now it is,
regardless of recording state.