udp: don't abort on recoverable errors; count lost packets
Certain recoverable UDP errors should not be treated as 'fatal', such
as OoO or duplicate packets.

Also, do not bother recording the number of 'short' packets - I am not
sure if it is even possible to get a short packet.  We will check for
and log them, but otherwise not care.

Keep a count of lost packets instead, which is much more likely, and
much more useful to know.

Change the status information label text to Magenta on the 'O'ptions
screen - easier to read in an overlay.