utModPlusOne(): remove and just use mod()
This function was designed to handle the cases of indexing into an
array with a starting index of 1, since RATFOR arrays always started
at 1, not 0 like the rest of the planet.

We don't need this anymore, and indeed this function has been modified
over the years to essentially do a modulus operation (x % y), but in a
much more complicated way.  So, just get rid of it and just use
standard modulus functionality.

Also, fix compile error in utDeleteBlanks().