team: allow disabling of teams
This commit allows teams to be disabled.  When the planets are
initialized, if a planet is found that is owned by a team, that team
is "enabled".  Any teams for which there are no team planets defined
is "disabled".

This affects what team is randomly selected for a new user, as well as
what team you can select yourself if the SwitchTeams server option is
enabled.