prm: stringify the implementation, touched on a lot of connecting code
Re-implement the prompting mechanism to use std::string for buffers
instead of random buffer pointers.  A lot of connecting code also had
to be modified (for the better!) to support this.

In addition, there were a few more std::string overloads added for
certain utility functions, like utIsDigits().

Also removed the dependence on GL headers so we can use prm_t
generically.  This should allow us to use this for curses (conqoper)
in the future, replacing the whole cdget/cdgetx mess.  Hopefully.