Rename checkuname() to uaValidateUsername(), adjust valid username characters
We use crypt for encrypting passwords (yeah, I know...) - it cannot
handle certain characters used as a salt (-_) which are (were) valid
characters in usernames.  So, we change it so that only [A-Za-z0-9] are
valid characters in user names now.  No more "-_" chars allowed.

Also, detect NULL returns from crypt and send BADPASSWD when it happens.