ui: Create uiGL and uiCU (OpenGL and Curses) implementation files
The idea here is to try to streamline the user interface calls (at
least those that make sense) into a "standard" API.  This will make it
possible to create more generic code (like the infoship/infoplanet
stuff) in the future that works the same way regardless of the
underlying rendering technology in use.

This was the original purpose of the UI stuff, but hasn't been
utilized much in the past.  Where it makes sense, we will try to
create a more generic interface to some of this stuff, and in the end,
reduce code duplication (like the info stuff mentioned above).

Created new uiCU.cxx (curses) and uiGL.cxx (OpenGL) implementation
files.  Currently, only uiPutMsg() (replacing cp_putmsg() in GL, and
mcuPutMsg() in curses), and uiMoveCursor() (no-op in GL, cdmove() call
in curses) are implemented.  I expect more to be coming.