Rename setWarp() to cqsSetWarp() and move into the sound library
All this function does is change the engine sounds depending on warp,
and delta warp.  So move it out of GL.cxx and move into it's own
cqsSetWarp.cxx file.  Remove call to cqsSetWarp() from hud functions so
that the hud data and setters are now UI agnostic.

hud.h, hud.cxx should now be usable in curses (conqoper), and we have
moved them into the main conquest library from the UiGL library.