Implement the shiptypes protocol
The shiptype sections in conqinitrc are still ignored.  This
commit just adds the necessary protocol support, but does not let you
actually change shiptype data.

It also adds a new cbLimits.maxShiptypes() which is stored in
recording files, and is sent to the client in the server hello packet
along with the other limits.

Shiptype data is also sent to the client, and the client uses that
data (decodes and stores it in the common block for use).

There's just no way in the server yet to allow editing existing
shiptypes or creating new ones.

Perhaps in the distant future...