Authenticate: detect when a user just hits return and clean up
When a user just hits return when prompted for a user name, the client
sends a CP_COMMAND/CPCMD_DISCONNECT and exits.  This is fine.

The server would not handle this as cleanly as it should however - it
would issue some error messages and then eventually exit when the
network connection disappeared shortly after.

Now, check for this case specifically, and don't be so whiney when it
happens.